pallet_registry::pallet

Trait Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Currency: Mutate<Self::AccountId> + MutateHold<Self::AccountId, Reason = Self::RuntimeHoldReason>;
    type WeightInfo: WeightInfo;
    type CanRegister: CanRegisterIdentity<Self::AccountId>;
    type MaxAdditionalFields: Get<u32>;
    type InitialDeposit: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type FieldDeposit: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type RuntimeHoldReason: From<HoldReason>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Currency: Mutate<Self::AccountId> + MutateHold<Self::AccountId, Reason = Self::RuntimeHoldReason>

Currency type that will be used to place deposits on neurons

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type CanRegister: CanRegisterIdentity<Self::AccountId>

Interface to allow other pallets to control who can register identities

type MaxAdditionalFields: Get<u32>

Configuration fields Maximum user-configured additional fields

type InitialDeposit: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

The amount held on deposit for a registered identity

type FieldDeposit: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

The amount held on deposit per additional field for a registered identity.

type RuntimeHoldReason: From<HoldReason>

Reasons for putting funds on hold.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.